Refactored streamlines API and added unit tests for LazyTractogram
