+++ nibabel/streamlines/tests/test_tractogram.py
@@ -1,4 +1,3 @@
-import os
 import unittest
 import numpy as np
 import warnings
@@ -6,202 +5,12 @@ import warnings
 from nibabel.testing import assert_arrays_equal, isiterable
 from nibabel.testing import suppress_warnings, clear_and_catch_warnings
 from nose.tools import assert_equal, assert_raises, assert_true
-from numpy.testing import assert_array_equal, assert_array_almost_equal
-from nibabel.externals.six.moves import zip, zip_longest
+from numpy.testing import assert_array_equal
+from nibabel.externals.six.moves import zip
 
-from .. import base_format
-from ..base_format import CompactList
-from ..base_format import TractogramItem, Tractogram, LazyTractogram
-from ..base_format import UsageWarning
-
-DATA_PATH = os.path.join(os.path.dirname(__file__), 'data')
-
-
-class TestCompactList(unittest.TestCase):
-
-    def setUp(self):
-        rng = np.random.RandomState(42)
-        self.data = [rng.rand(rng.randint(10, 50), 3) for _ in range(10)]
-        self.lengths = map(len, self.data)
-        self.clist = CompactList(self.data)
-
-    def test_creating_empty_compactlist(self):
-        clist = CompactList()
-        assert_equal(len(clist), 0)
-        assert_equal(len(clist._offsets), 0)
-        assert_equal(len(clist._lengths), 0)
-        assert_true(clist._data is None)
-        assert_true(clist.shape is None)
-
-    def test_creating_compactlist_from_list(self):
-        rng = np.random.RandomState(42)
-        data = [rng.rand(rng.randint(10, 50), 3) for _ in range(10)]
-        lengths = map(len, data)
-
-        clist = CompactList(data)
-        assert_equal(len(clist), len(data))
-        assert_equal(len(clist._offsets), len(data))
-        assert_equal(len(clist._lengths), len(data))
-        assert_equal(clist._data.shape[0], sum(lengths))
-        assert_equal(clist._data.shape[1], 3)
-        assert_equal(clist._offsets, [0] + np.cumsum(lengths)[:-1].tolist())
-        assert_equal(clist._lengths, lengths)
-        assert_equal(clist.shape, data[0].shape[1:])
-
-        # Empty list
-        clist = CompactList([])
-        assert_equal(len(clist), 0)
-        assert_equal(len(clist._offsets), 0)
-        assert_equal(len(clist._lengths), 0)
-        assert_true(clist._data is None)
-        assert_true(clist.shape is None)
-
-    def test_creating_compactlist_from_generator(self):
-        rng = np.random.RandomState(42)
-        data = [rng.rand(rng.randint(10, 50), 3) for _ in range(10)]
-        lengths = map(len, data)
-
-        gen = (e for e in data)
-        clist = CompactList(gen)
-        assert_equal(len(clist), len(data))
-        assert_equal(len(clist._offsets), len(data))
-        assert_equal(len(clist._lengths), len(data))
-        assert_equal(clist._data.shape[0], sum(lengths))
-        assert_equal(clist._data.shape[1], 3)
-        assert_equal(clist._offsets, [0] + np.cumsum(lengths)[:-1].tolist())
-        assert_equal(clist._lengths, lengths)
-        assert_equal(clist.shape, data[0].shape[1:])
-
-        # Already consumed generator
-        clist = CompactList(gen)
-        assert_equal(len(clist), 0)
-        assert_equal(len(clist._offsets), 0)
-        assert_equal(len(clist._lengths), 0)
-        assert_true(clist._data is None)
-        assert_true(clist.shape is None)
-
-    def test_creating_compactlist_from_compact_list(self):
-        rng = np.random.RandomState(42)
-        data = [rng.rand(rng.randint(10, 50), 3) for _ in range(10)]
-        lengths = map(len, data)
-
-        clist = CompactList(data)
-        clist2 = CompactList(clist)
-        assert_equal(len(clist2), len(data))
-        assert_equal(len(clist2._offsets), len(data))
-        assert_equal(len(clist2._lengths), len(data))
-        assert_equal(clist2._data.shape[0], sum(lengths))
-        assert_equal(clist2._data.shape[1], 3)
-        assert_equal(clist2._offsets, [0] + np.cumsum(lengths)[:-1].tolist())
-        assert_equal(clist2._lengths, lengths)
-        assert_equal(clist2.shape, data[0].shape[1:])
-
-    def test_compactlist_iter(self):
-        for e, d in zip(self.clist, self.data):
-            assert_array_equal(e, d)
-
-    def test_compactlist_copy(self):
-        clist = self.clist.copy()
-        assert_array_equal(clist._data, self.clist._data)
-        assert_true(clist._data is not self.clist._data)
-        assert_array_equal(clist._offsets, self.clist._offsets)
-        assert_true(clist._offsets is not self.clist._offsets)
-        assert_array_equal(clist._lengths, self.clist._lengths)
-        assert_true(clist._lengths is not self.clist._lengths)
-
-        assert_equal(clist.shape, self.clist.shape)
-
-        # When taking a copy of a `CompactList` generated by slicing.
-        # Only needed data should be kept.
-        clist = self.clist[::2].copy()
-
-        assert_true(clist._data.shape[0] < self.clist._data.shape[0])
-        assert_true(len(clist) < len(self.clist))
-        assert_true(clist._data is not self.clist._data)
-
-    def test_compactlist_append(self):
-        # Maybe not necessary if `self.setUp` is always called before a
-        # test method, anyways create a copy just in case.
-        clist = self.clist.copy()
-
-        rng = np.random.RandomState(1234)
-        element = rng.rand(rng.randint(10, 50), *self.clist.shape)
-        clist.append(element)
-        assert_equal(len(clist), len(self.clist)+1)
-        assert_equal(clist._offsets[-1], len(self.clist._data))
-        assert_equal(clist._lengths[-1], len(element))
-        assert_array_equal(clist._data[-len(element):], element)
-
-        # Append with different shape.
-        element = rng.rand(rng.randint(10, 50), 42)
-        assert_raises(ValueError, clist.append, element)
-
-        # Append to an empty CompactList.
-        clist = CompactList()
-        rng = np.random.RandomState(1234)
-        shape = (2, 3, 4)
-        element = rng.rand(rng.randint(10, 50), *shape)
-        clist.append(element)
-
-        assert_equal(len(clist), 1)
-        assert_equal(clist._offsets[-1], 0)
-        assert_equal(clist._lengths[-1], len(element))
-        assert_array_equal(clist._data, element)
-        assert_equal(clist.shape, shape)
-
-    def test_compactlist_extend(self):
-        # Maybe not necessary if `self.setUp` is always called before a
-        # test method, anyways create a copy just in case.
-        clist = self.clist.copy()
-
-        rng = np.random.RandomState(1234)
-        shape = self.clist.shape
-        new_data = [rng.rand(rng.randint(10, 50), *shape) for _ in range(5)]
-        lengths = map(len, new_data)
-        clist.extend(new_data)
-        assert_equal(len(clist), len(self.clist)+len(new_data))
-        assert_array_equal(clist._offsets[-len(new_data):],
-                           len(self.clist._data) + np.cumsum([0] + lengths[:-1]))
-
-        assert_equal(clist._lengths[-len(new_data):], lengths)
-        assert_array_equal(clist._data[-sum(lengths):],
-                           np.concatenate(new_data, axis=0))
-
-        # Extend with another `CompactList` object.
-        clist = self.clist.copy()
-        new_data = CompactList(new_data)
-        clist.extend(new_data)
-        assert_equal(len(clist), len(self.clist)+len(new_data))
-        assert_array_equal(clist._offsets[-len(new_data):],
-                           len(self.clist._data) + np.cumsum([0] + lengths[:-1]))
-
-        assert_equal(clist._lengths[-len(new_data):], lengths)
-        assert_array_equal(clist._data[-sum(lengths):], new_data._data)
-
-    def test_compactlist_getitem(self):
-        # Get one item
-        for i, e in enumerate(self.clist):
-            assert_array_equal(self.clist[i], e)
-
-        # Get multiple items (this will create a view).
-        clist_view = self.clist[range(len(self.clist))]
-        assert_true(clist_view is not self.clist)
-        assert_true(clist_view._data is self.clist._data)
-        assert_true(clist_view._offsets is not self.clist._offsets)
-        assert_true(clist_view._lengths is not self.clist._lengths)
-        assert_array_equal(clist_view._offsets, self.clist._offsets)
-        assert_array_equal(clist_view._lengths, self.clist._lengths)
-        for e1, e2 in zip_longest(clist_view, self.clist):
-            assert_array_equal(e1, e2)
-
-        # Get slice (this will create a view).
-        clist_view = self.clist[::2]
-        assert_true(clist_view is not self.clist)
-        assert_true(clist_view._data is self.clist._data)
-        assert_array_equal(clist_view._offsets, self.clist._offsets[::2])
-        assert_array_equal(clist_view._lengths, self.clist._lengths[::2])
-        for i, e in enumerate(clist_view):
-            assert_array_equal(e, self.clist[i*2])
+from .. import tractogram as module_tractogram
+from ..tractogram import UsageWarning
+from ..tractogram import TractogramItem, Tractogram, LazyTractogram
 
 
 class TestTractogramItem(unittest.TestCase):
@@ -247,7 +56,7 @@ class TestTractogram(unittest.TestCase):
                                     [0, 0, 1],
                                     [1, 0, 0]], dtype="f4")
 
-        self.nb_tractogram = len(self.streamlines)
+        self.nb_streamlines = len(self.streamlines)
 
     def test_tractogram_creation(self):
         # Create an empty tractogram.
@@ -379,10 +188,6 @@ class TestTractogram(unittest.TestCase):
 class TestLazyTractogram(unittest.TestCase):
 
     def setUp(self):
-        self.empty_trk_filename = os.path.join(DATA_PATH, "empty.trk")
-        self.simple_trk_filename = os.path.join(DATA_PATH, "simple.trk")
-        self.complex_trk_filename = os.path.join(DATA_PATH, "complex.trk")
-
         self.streamlines = [np.arange(1*3, dtype="f4").reshape((1, 3)),
                             np.arange(2*3, dtype="f4").reshape((2, 3)),
                             np.arange(5*3, dtype="f4").reshape((5, 3))]
@@ -391,71 +196,81 @@ class TestLazyTractogram(unittest.TestCase):
                        np.array([(0, 1, 0)]*2, dtype="f4"),
                        np.array([(0, 0, 1)]*5, dtype="f4")]
 
-        self.mean_curvature_torsion = [np.array([1.11, 1.22], dtype="f4"),
-                                       np.array([2.11, 2.22], dtype="f4"),
-                                       np.array([3.11, 3.22], dtype="f4")]
+        self.mean_curvature = np.array([1.11, 2.11, 3.11], dtype="f4")
+        self.mean_color = np.array([[0, 1, 0],
+                                    [0, 0, 1],
+                                    [1, 0, 0]], dtype="f4")
 
         self.nb_streamlines = len(self.streamlines)
-        self.nb_scalars_per_point = self.colors[0].shape[1]
-        self.nb_properties_per_streamline = len(self.mean_curvature_torsion[0])
 
     def test_lazy_tractogram_creation(self):
         # To create tractogram from arrays use `Tractogram`.
         assert_raises(TypeError, LazyTractogram, self.streamlines)
 
-        # Points, scalars and properties
+        # Streamlines and other data as generators
         streamlines = (x for x in self.streamlines)
-        scalars = (x for x in self.colors)
-        properties = (x for x in self.mean_curvature_torsion)
+        data_per_point = {"colors": (x for x in self.colors)}
+        data_per_streamline = {'mean_curv': (x for x in self.mean_curvature),
+                               'mean_color': (x for x in self.mean_color)}
 
-        # Creating LazyTractogram from generators is not allowed as
-        # generators get exhausted and are not reusable unline coroutines.
+        # Creating LazyTractogram with generators is not allowed as
+        # generators get exhausted and are not reusable unlike coroutines.
         assert_raises(TypeError, LazyTractogram, streamlines)
-        assert_raises(TypeError, LazyTractogram, self.streamlines, scalars)
-        assert_raises(TypeError, LazyTractogram, properties_func=properties)
+        assert_raises(TypeError, LazyTractogram,
+                      data_per_streamline=data_per_streamline)
+        assert_raises(TypeError, LazyTractogram, self.streamlines,
+                      data_per_point=data_per_point)
 
         # Empty `LazyTractogram`
         tractogram = LazyTractogram()
-        with suppress_warnings():
-            assert_equal(len(tractogram), 0)
+        assert_true(isiterable(tractogram))
+        assert_equal(len(tractogram), 0)
         assert_arrays_equal(tractogram.streamlines, [])
-        assert_arrays_equal(tractogram.scalars, [])
-        assert_arrays_equal(tractogram.properties, [])
+        assert_equal(tractogram.data_per_point, {})
+        assert_equal(tractogram.data_per_streamline, {})
 
-        # Check if we can iterate through the tractogram.
-        for streamline in tractogram:
-            pass
-
-        # Points, scalars and properties
+        # Create tractogram with streamlines and other data
         streamlines = lambda: (x for x in self.streamlines)
-        scalars = lambda: (x for x in self.colors)
-        properties = lambda: (x for x in self.mean_curvature_torsion)
+        data_per_point = {"colors": lambda: (x for x in self.colors)}
+        data_per_streamline = {'mean_curv': lambda: (x for x in self.mean_curvature),
+                               'mean_color': lambda: (x for x in self.mean_color)}
 
-        tractogram = LazyTractogram(streamlines, scalars, properties)
-        with suppress_warnings():
-            assert_equal(len(tractogram), self.nb_streamlines)
+        tractogram = LazyTractogram(streamlines,
+                                    data_per_streamline=data_per_streamline,
+                                    data_per_point=data_per_point)
+
+        assert_true(isiterable(tractogram))
+        assert_equal(len(tractogram), self.nb_streamlines)
 
         # Coroutines get re-called and creates new iterators.
-        assert_arrays_equal(tractogram.streamlines, self.streamlines)
-        assert_arrays_equal(tractogram.scalars, self.colors)
-        assert_arrays_equal(tractogram.properties, self.mean_curvature_torsion)
-        assert_arrays_equal(tractogram.streamlines, self.streamlines)
-        assert_arrays_equal(tractogram.scalars, self.colors)
-        assert_arrays_equal(tractogram.properties, self.mean_curvature_torsion)
+        for i in range(2):
+            assert_arrays_equal(tractogram.streamlines, self.streamlines)
+            assert_arrays_equal(tractogram.data_per_streamline['mean_curv'],
+                                self.mean_curvature)
+            assert_arrays_equal(tractogram.data_per_streamline['mean_color'],
+                                self.mean_color)
+            assert_arrays_equal(tractogram.data_per_point['colors'],
+                                self.colors)
 
         # Create `LazyTractogram` from a coroutine yielding 3-tuples
-        data = lambda: (x for x in zip(self.streamlines, self.colors, self.mean_curvature_torsion))
+        def _data_gen():
+            for d in zip(self.streamlines, self.colors,
+                         self.mean_curvature, self.mean_color):
+                data_for_points = {'colors': d[1]}
+                data_for_streamline = {'mean_curv': d[2],
+                                       'mean_color': d[3]}
+                yield TractogramItem(d[0], data_for_streamline, data_for_points)
 
-        tractogram = LazyTractogram.create_from_data(data)
-        with suppress_warnings():
-            assert_equal(len(tractogram), self.nb_streamlines)
+        tractogram = LazyTractogram.create_from(_data_gen)
+        assert_true(isiterable(tractogram))
+        assert_equal(len(tractogram), self.nb_streamlines)
         assert_arrays_equal(tractogram.streamlines, self.streamlines)
-        assert_arrays_equal(tractogram.scalars, self.colors)
-        assert_arrays_equal(tractogram.properties, self.mean_curvature_torsion)
-
-        # Check if we can iterate through the tractogram.
-        for streamline in tractogram:
-            pass
+        assert_arrays_equal(tractogram.data_per_streamline['mean_curv'],
+                            self.mean_curvature)
+        assert_arrays_equal(tractogram.data_per_streamline['mean_color'],
+                            self.mean_color)
+        assert_arrays_equal(tractogram.data_per_point['colors'],
+                            self.colors)
 
     def test_lazy_tractogram_indexing(self):
         streamlines = lambda: (x for x in self.streamlines)
@@ -473,7 +288,8 @@ class TestLazyTractogram(unittest.TestCase):
 
             return list(zip(self.streamlines[idx], self.colors[idx]))
 
-        tractogram = LazyTractogram(streamlines, scalars, properties, getitem_without_properties)
+        tractogram = LazyTractogram(streamlines, scalars, properties,
+                                    getitem_without_properties)
         streamlines, scalars = tractogram[0]
         assert_array_equal(streamlines, self.streamlines[0])
         assert_array_equal(scalars, self.colors[0])
@@ -491,7 +307,7 @@ class TestLazyTractogram(unittest.TestCase):
         scalars = lambda: (x for x in self.colors)
         properties = lambda: (x for x in self.mean_curvature_torsion)
 
-        with clear_and_catch_warnings(record=True, modules=[base_format]) as w:
+        with clear_and_catch_warnings(record=True, modules=[module_tractogram]) as w:
             warnings.simplefilter("always")  # Always trigger warnings.
 
             # Calling `len` will create new generators each time.
@@ -510,19 +326,20 @@ class TestLazyTractogram(unittest.TestCase):
             assert_equal(len(tractogram), self.nb_streamlines)
             assert_equal(len(w), 2)
 
-        with clear_and_catch_warnings(record=True, modules=[base_format]) as w:
+        with clear_and_catch_warnings(record=True, modules=[module_tractogram]) as w:
             # Once we iterated through the tractogram, we know the length.
             tractogram = LazyTractogram(streamlines, scalars, properties)
             assert_true(tractogram.header.nb_streamlines is None)
             for streamline in tractogram:
                 pass
 
-            assert_equal(tractogram.header.nb_streamlines, len(self.streamlines))
+            assert_equal(tractogram.header.nb_streamlines,
+                         len(self.streamlines))
             # This should *not* produce a warning.
             assert_equal(len(tractogram), len(self.streamlines))
             assert_equal(len(w), 0)
 
-        with clear_and_catch_warnings(record=True, modules=[base_format]) as w:
+        with clear_and_catch_warnings(record=True, modules=[module_tractogram]) as w:
             # It first checks if number of tractogram is in the header.
             tractogram = LazyTractogram(streamlines, scalars, properties)
             tractogram.header.nb_streamlines = 1234
